batch: Group merge policy modules
The batch package has bounded matching, realization, and ownership
dependencies while merge constraints and placement policy remain spread
across flat modules.

Following one merge decision requires contributors to distinguish baseline
edits, presence placement, absence handling, candidate enumeration, and
validation from unrelated batch concerns.

This commit moves the connected merge policy beneath merge, shortens
repeated candidate names, and keeps discard as a separate batch-level
operation.

Later commits will migrate operation candidates, source handling, storage,
and display before the temporary merge imports are removed.
