+++ src/git_stage_batch/batch/merge/__init__.py
@@ -1,0 +1,16 @@
+"""Placement of saved ownership into current file content."""
+
+from .merge import (
+    _merge_batch_line_chunks,
+    can_merge_batch_from_line_sequences,
+    enumerate_merge_batch_candidates_from_line_sequences,
+    merge_batch_from_line_sequences_as_buffer,
+)
+
+
+__all__ = [
+    "_merge_batch_line_chunks",
+    "can_merge_batch_from_line_sequences",
+    "enumerate_merge_batch_candidates_from_line_sequences",
+    "merge_batch_from_line_sequences_as_buffer",
+]
